rename osx files to macos
