+++ MANIFEST.in
@@ -28,7 +28,7 @@ include psutil/_exceptions.py
 include psutil/_psaix.py
 include psutil/_psbsd.py
 include psutil/_pslinux.py
-include psutil/_psosx.py
+include psutil/_psmacos.py
 include psutil/_psposix.py
 include psutil/_pssunos.py
 include psutil/_psutil_aix.c
@@ -36,7 +36,7 @@ include psutil/_psutil_bsd.c
 include psutil/_psutil_common.c
 include psutil/_psutil_common.h
 include psutil/_psutil_linux.c
-include psutil/_psutil_osx.c
+include psutil/_psutil_macos.c
 include psutil/_psutil_posix.c
 include psutil/_psutil_posix.h
 include psutil/_psutil_sunos.c
@@ -55,14 +55,16 @@ include psutil/arch/freebsd/specific.c
 include psutil/arch/freebsd/specific.h
 include psutil/arch/freebsd/sys_socks.c
 include psutil/arch/freebsd/sys_socks.h
+include psutil/arch/macos/process_info.c
+include psutil/arch/macos/process_info.h
+include psutil/arch/macos/smc.c
+include psutil/arch/macos/smc.h
 include psutil/arch/netbsd/socks.c
 include psutil/arch/netbsd/socks.h
 include psutil/arch/netbsd/specific.c
 include psutil/arch/netbsd/specific.h
 include psutil/arch/openbsd/specific.c
 include psutil/arch/openbsd/specific.h
-include psutil/arch/osx/process_info.c
-include psutil/arch/osx/process_info.h
 include psutil/arch/solaris/environ.c
 include psutil/arch/solaris/environ.h
 include psutil/arch/solaris/v10/ifaddrs.c
@@ -86,9 +88,9 @@ include psutil/tests/test_bsd.py
 include psutil/tests/test_connections.py
 include psutil/tests/test_contracts.py
 include psutil/tests/test_linux.py
+include psutil/tests/test_macos.py
 include psutil/tests/test_memory_leaks.py
 include psutil/tests/test_misc.py
-include psutil/tests/test_osx.py
 include psutil/tests/test_posix.py
 include psutil/tests/test_process.py
 include psutil/tests/test_sunos.py
