+++ psutil/arch/macos/process_info.c
@@ -4,7 +4,7 @@
  * found in the LICENSE file.
  *
  * Helper functions related to fetching process information.
- * Used by _psutil_osx module methods.
+ * Used by _psutil_macos module methods.
  */
 
 
@@ -142,7 +142,7 @@ psutil_get_cmdline(long pid) {
     mib[2] = (pid_t)pid;
     if (sysctl(mib, 3, procargs, &argmax, NULL, 0) < 0) {
         // In case of zombie process we'll get EINVAL. We translate it
-        // to NSP and _psosx.py will translate it to ZP.
+        // to NSP and _psmacos.py will translate it to ZP.
         if ((errno == EINVAL) && (psutil_pid_exists(pid)))
             NoSuchProcess("");
         else
@@ -236,7 +236,7 @@ psutil_get_environ(long pid) {
     mib[2] = (pid_t)pid;
     if (sysctl(mib, 3, procargs, &argmax, NULL, 0) < 0) {
         // In case of zombie process we'll get EINVAL. We translate it
-        // to NSP and _psosx.py will translate it to ZP.
+        // to NSP and _psmacos.py will translate it to ZP.
         if ((errno == EINVAL) && (psutil_pid_exists(pid)))
             NoSuchProcess("");
         else
