+++ psutil/arch/macos/smc.c
