+++ psutil/arch/macos/smc.h
