+++ psutil/tests/test_macos.py
