+++ setup.py
@@ -148,11 +148,11 @@ if WINDOWS:
 elif MACOS:
     macros.append(("PSUTIL_MACOS", 1))
     ext = Extension(
-        'psutil._psutil_osx',
+        'psutil._psutil_macos',
         sources=sources + [
-            'psutil/_psutil_osx.c',
-            'psutil/arch/osx/process_info.c',
-            'psutil/arch/osx/smc.c',
+            'psutil/_psutil_macos.c',
+            'psutil/arch/macos/process_info.c',
+            'psutil/arch/macos/smc.c',
         ],
         define_macros=macros,
         extra_link_args=[
